refactor: move post rate limiter
