+++ src/lib/rate-limiter/rate-limiter-post.ts
@@ -1,10 +1,10 @@
 import config from 'config';
 import { RateLimiterRedis, RateLimiterRes } from 'rate-limiter-flexible';
 import requestIp from 'request-ip';
-import { getPersistentRedisClient } from './redis/persistent-client.js';
+import { getPersistentRedisClient } from '../redis/persistent-client.js';
 import createHttpError from 'http-errors';
-import type { ExtendedContext } from '../types.js';
-import { credits } from './credits.js';
+import type { ExtendedContext } from '../../types.js';
+import { credits } from '../credits.js';
 
 const redisClient = getPersistentRedisClient();
